Working Version.

+++ src/team5115/subsystems/Drivetrain.java
@@ -118,8 +118,7 @@ public class Drivetrain implements DriveBase {
     public double getAvgSpd() {
         double rightSpd = frontRight.getSelectedSensorVelocity();
         double leftSpd = -backLeft.getSelectedSensorVelocity();
-
-        final double wheelSpd = ((rightSpd + leftSpd) / 2) * 1.53846 * Math.PI / 4090;
+        final double wheelSpd = ((rightSpd + leftSpd) / (double) 2) * 1.53846 * Math.PI / 4090;
         System.out.println("Wheel Speeds = " + wheelSpd);
         return wheelSpd;
     }
